move CRUD logic of layouts tolayoutrepository
